Working on camera frustum culling

+++ src/components/Destroy.h
@@ -2,7 +2,7 @@
 #define VOX_DESTROY_H
 
 struct Destroy {
-    float time{0};
+    double time{0};
 };
 
 #endif //VOX_DESTROY_H
